backend/requirements.txt
langchain==0.2.16
langchain-groq
langchain-core==0.2.38
langgraph==0.2.28
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
python-dotenv==1.0.1
pydantic==2.7.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
scipy==1.13.1
